Replaced stat ineff with autocorr from emcee.
The statistical inneficiency consistently estimated garbage. I replaced it
with the autocorr cuz conceptually it is the exact same except there is no
weird block averaging. Im not sure I fully get the idea behind that method yet.
The autocorr applied sokals critera with the window function. Not 100 percent
sure on the current breakdown of how callbacks and stages work. Should they have
access to the kc object or not?